+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Allow adding more query types (still defaults to just IMDb and title queries)
   - Reduce number of requests when scraping needforseed.php
 - New torrent fields: `LastActive`, `LastReseedRequest`, `ReseedWaitingUsers`, `Trumpable`
+- `ptp-reseed`: Add client configuration section in ptpapi.conf
 ### Fixed
 - `ptp-reseed`: Fix rtorrent existence check